newfs_msdos: fix type of kern.maxphys
The type of the kern.maxphys sysctl OID is now ulong.  Change the
local variable type to match.

Reviewed by:	delphij, emaste
Sponsored by:	Dell EMC Isilon

(cherry picked from commit 9990450e17f79ca6e59866397b979a8a344cd5d4)

+++ sbin/newfs_msdos/mkfs_msdos.c
@@ -844,10 +844,10 @@ check_mounted(const char *fname, mode_t mode)
 static ssize_t
 getchunksize(void)
 {
-	static int chunksize;
+	static ssize_t chunksize;
 
 	if (chunksize != 0)
-		return ((ssize_t)chunksize);
+		return (chunksize);
 
 #ifdef	KERN_MAXPHYS
 	int mib[2];
@@ -876,7 +876,7 @@ getchunksize(void)
 	assert(powerof2(chunksize));
 	assert(chunksize > MAXBPS);
 
-	return ((ssize_t)chunksize);
+	return (chunksize);
 }
 
 /*
